Created a channel to use for all the different routes
I added a sample signup and login with routes, to help me with the route designs. It works well so far and interacts with sequelize and database.

It uses express. Has passport but has not been added it to the code yet.

signup Route and login route both have authentication and passwords are hashed enroute and when accessed for login purposes from database(using bcrypt.

Login route will give an error as it needs a dashboard to route the user to. We can discuss how we want to implement the dashboard and then I can have it route there.

I updated sequelize and database to allow hashed passwords